Added img on pdf; Added local date/time in Prepared by
Added the Crossfeed logo in upper right corner of generated PDF. Had to hardcode the base64 rendering of the image. Perhaps not the best way to do this but it works.

Added the local version of the date and time in the prepared by section to make human readable verses the GMT date/time.